+++ src/Guzzle/Http/Client.php
@@ -157,7 +157,7 @@ class Client extends AbstractHasDispatcher implements ClientInterface
 
     public function expandTemplate($template, array $variables = null)
     {
-        $expansionVars = $this->getConfig()->getAll();
+        $expansionVars = $this->getConfig()->toArray();
         if ($variables) {
             $expansionVars = $variables + $expansionVars;
         }